initial refactor to constants
